Fixed a deadlock in the Runner
If there was any communication error with the grader while a run was in
progress, the finished channel and the ChannelBuffer could end up in a
state where the goroutines in which they run got stuck waiting on each
other. This change fixes that and also adds a debugging web server to
inspect the status of the goroutines.