Better logging+retry management
Previously some runs would encounter an error and instead of immediately
retrying, they would wait 5 minutes until the timeout mechanism kicked
in. This change makes it so that any recoverable error will cause the
run to be requeued in order to retry immediately, and fatal errors will
cause it to fail permanently.

Logging was also improved a bit, since the logging context for a run now
always prints the run id.