Preload sandbox inputs
Pre-fetching input files makes contestant's programs use 3-30% less
user/wall time in I/O, so it seems like the right thing to do.

This change will first attempt to mmap all the file and touch one byte
in all pages to ensure it is all mapped into memory. If this fails, we
just sequentially read the whole file to ensure it's in the OS cache at
least.