Add a fairer ioLock
Runners tend to prefer preloading inputs rather than actually serving
runs. This is due to the combination of Go's sync.Mutex not being fair
and Go's scheduler preferring to keep running the current goroutine if
it is still ready. This change adds rsc's workaround to make the ioLock
a bit fairer.