Concat minijail error logs to the .err file
Now that minijail does not use syslog to log, it is super useful to
print out the minijail errors in the .err file. It cannot be done in a
single step because the contestant can overwrite the file, so create a
temporary file and append it to the original .err if it is non-empty.