Ephemeral grader
This change adds a somewhat experimental mode of the grader where there
does not need to be a backing problem to run something. This allows
users to try stuff out without needing to go through the hassle of
uploading a problem.

This of course also supports all the bells and whistles of a normal
problem, like libinteractive and custom validators.