Limit the number of concurrent CI requests
The CI requests can be pretty memory-intensive, since we are storing the
full request in memory. This change makes it such that there is at most
one such request in flight by moving the code that handles that to its
own goroutine, using a 128-deep channel to store incoming pre-lookup
requests.

This also exposes whatever errors happen during the asynchronous part of
the CI run in a new field.