Change the expected input format of -oneshot=run
This change makes it easier to run omegaup-runner in oneshot mode by not
requiring the .sha1 file with the hashes of all the files in the input
to be there already. Now the only requirement for -oneshot=run is that
the input directory should contain a settings.json file, which checkouts
of problems already have.