Move the CI from Travis to GitHub Actions
This change makes the CI to be run by GitHub Actions instead of Travis.
This is a lot less flaky and a tad bit faster.

This also automates the release process and builds the artifacts so that
I don't have to run a weird script on my machine every time.