Make the RunPostProcessor more robust
This change makes the RunPostProcessor a member of the QueueManager
instead of the InflightMonitor. This is because the InflightMonitor can
be removed from the RunContext before calling Close(). This then causes
the run to then not invoke the RunPostProcessor, which in turn makes
runs to be in this state to be permanently stuck in the 'new' state.