Refactor the queue management logic
This change splits the interface to add runs in two flavors: one that
returns a RunWaitHandle and one that does not. This is because it is
kind of weird to ignore one of the return values sometimes.

It also clarifies that the blocking behavior is here to stay, for all
the non-ephemeral runs.