Prevent panics in some JEs
There were some cases where a run fails so hard that it needs a special
RunInfo with a JE verdict. The problem is that when we try to save this
into the database, it panics.

This change prevents that panic by creating the object with the correct
non-nil fields. The version of go-base was also bumped to prevent a
panic _even_ if it is nil.